requirements.txt
datetime
python-dotenv
boto3
rich
torch
torchvision
torchaudio
numpy
typing
pandas
botocore